Add button to check in toast to edit attendee details
