+++ frontend/src/pages/ManageMeetupAttendeesPage.tsx
@@ -1,14 +1,5 @@
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import { Input } from '@/components/ui/input';
 import { Spinner } from '@/components/ui/spinner';
 import { useDisclosure } from '@/hooks/useDisclosure';
 import { type TicketInfo } from '@keebmeet/shared';
@@ -16,15 +7,11 @@ import dayjs from 'dayjs';
 import { useState, type ReactNode } from 'react';
 import { FiEdit2 } from 'react-icons/fi';
 import { useParams } from 'react-router-dom';
-import { toast } from 'sonner';
+import { AttendeeDetailsDialog } from '../components/AttendeeDetailsDialog';
 import { DataTable, type DataTableColumn } from '../components/DataTable';
 import { ExpandableCard } from '../components/ExpandableCard';
 import { useGetMeetupQuery } from '../store/meetupSlice';
-import {
-  useEditAttendeeMutation,
-  useGetMeetupAttendeesQuery,
-  useRefundAttendeeMutation,
-} from '../store/organizerSlice';
+import { useGetMeetupAttendeesQuery } from '../store/organizerSlice';
 
 const paymentStatusBadge = (
   status: TicketInfo['payment_status']
@@ -69,47 +56,17 @@ const ManageMeetupAttendeesPage = (): ReactNode => {
     { skip: meetup == null }
   );
 
-  const [editAttendee, { isLoading: isSaving }] = useEditAttendeeMutation();
-  const [refundAttendee, { isLoading: isRefunding }] =
-    useRefundAttendeeMutation();
   const { isOpen, onOpen, onClose } = useDisclosure();
 
   const [viewing, setViewing] = useState<TicketInfo | null>(null);
-  const [isEditing, setIsEditing] = useState<boolean>(false);
-  const [raffleEntries, setRaffleEntries] = useState<string>('');
-  const [refundConfirmOpen, setRefundConfirmOpen] = useState(false);
 
   const isPaidMeetup =
     meetup?.ticket_types?.some((type) => type.price_cents > 0) ?? false;
 
   const openDialog = (attendee: TicketInfo): void => {
     setViewing(attendee);
-    setRaffleEntries(String(attendee.raffle_entries));
-    setIsEditing(false);
     onOpen();
   };
-
-  const closeDialog = (): void => {
-    setViewing(null);
-    setIsEditing(false);
-    onClose();
-  };
-
-  const startEditing = (): void => {
-    if (viewing == null) return;
-    setRaffleEntries(String(viewing.raffle_entries));
-    setIsEditing(true);
-  };
-
-  const cancelEditing = (): void => {
-    if (viewing != null) {
-      setRaffleEntries(String(viewing.raffle_entries));
-    }
-    setIsEditing(false);
-  };
-
-  const entries = parseInt(raffleEntries, 10);
-  const canSave = Number.isInteger(entries) && entries >= 0;
 
   if (isLoading) {
     return (
@@ -118,52 +75,6 @@ const ManageMeetupAttendeesPage = (): ReactNode => {
       </div>
     );
   }
-
-  const handleSave = (): void => {
-    if (viewing == null || !canSave) return;
-
-    void (async () => {
-      const result = await editAttendee({
-        ticketId: viewing.id,
-        payload: {
-          raffle_entries: entries,
-        },
-      });
-
-      if ('error' in result) {
-        toast.error('Error', {
-          description: `Could not update ${viewing.ticket_holder_display_name}`,
-        });
-      } else {
-        toast.success('Success', {
-          description: `${viewing.ticket_holder_display_name} updated`,
-        });
-        // Reflect the saved value immediately and return to view mode.
-        setViewing({ ...viewing, raffle_entries: entries });
-        setIsEditing(false);
-      }
-    })();
-  };
-
-  const handleRefund = (): void => {
-    if (viewing == null) return;
-
-    void (async () => {
-      const result = await refundAttendee(viewing.id);
-
-      if ('error' in result) {
-        toast.error('Refund failed', {
-          description: `Could not refund ${viewing.ticket_holder_display_name}`,
-        });
-      } else {
-        toast.success('Ticket refunded', {
-          description: `${viewing.ticket_holder_display_name} was refunded`,
-        });
-        setViewing({ ...viewing, payment_status: 'refunded' });
-        setRefundConfirmOpen(false);
-      }
-    })();
-  };
 
   const columns: Array<DataTableColumn<TicketInfo>> = [
     {
@@ -293,141 +204,17 @@ const ManageMeetupAttendeesPage = (): ReactNode => {
         )}
       />
 
-      <Dialog
+      <AttendeeDetailsDialog
+        key={viewing?.id}
+        attendee={viewing}
         open={isOpen}
         onOpenChange={(open) => {
           if (!open) {
-            closeDialog();
+            onClose();
           }
         }}
-      >
-        <DialogContent showCloseButton={false}>
-          <DialogHeader>
-            <div className="flex items-center justify-between gap-2">
-              <DialogTitle>Attendee details</DialogTitle>
-              {!isEditing ? (
-                <Button
-                  variant="ghost"
-                  aria-label="Edit attendee"
-                  onClick={startEditing}
-                >
-                  <FiEdit2 />
-                  Edit
-                </Button>
-              ) : null}
-            </div>
-          </DialogHeader>
-          {viewing != null ? (
-            <dl className="grid grid-cols-2 items-center gap-x-4 gap-y-3 text-sm">
-              <dt className="text-muted-foreground">Display Name</dt>
-              <dd className="text-right">
-                {viewing.ticket_holder_display_name}
-              </dd>
-
-              {isPaidMeetup ? (
-                <>
-                  <dt className="text-muted-foreground">Status</dt>
-                  <dd className="ml-auto">
-                    {paymentStatusBadge(viewing.payment_status) ?? '—'}
-                  </dd>
-                </>
-              ) : null}
-
-              <dt className="text-muted-foreground">First Name</dt>
-              <dd className="text-right">
-                {orDash(viewing.ticket_holder_first_name)}
-              </dd>
-
-              <dt className="text-muted-foreground">Last Name</dt>
-              <dd className="text-right">
-                {orDash(viewing.ticket_holder_last_name)}
-              </dd>
-
-              <dt className="text-muted-foreground">Raffle Entries</dt>
-              <dd className="text-right">
-                {isEditing ? (
-                  <Input
-                    type="number"
-                    min={0}
-                    className="ml-auto w-24 text-right"
-                    value={raffleEntries}
-                    onChange={(e) => {
-                      setRaffleEntries(e.target.value);
-                    }}
-                  />
-                ) : (
-                  viewing.raffle_entries
-                )}
-              </dd>
-
-              <dt className="text-muted-foreground">Raffle Wins</dt>
-              <dd className="text-right">{viewing.raffle_wins}</dd>
-
-              <dt className="text-muted-foreground">Signed Up</dt>
-              <dd className="text-right">
-                {dayjs(viewing.created_at).format('M/D/YY hh:mm A')}
-              </dd>
-            </dl>
-          ) : null}
-          <DialogFooter>
-            {isEditing ? (
-              <>
-                <Button variant="outline" onClick={cancelEditing}>
-                  Cancel
-                </Button>
-                <Button disabled={!canSave || isSaving} onClick={handleSave}>
-                  Save
-                  {isSaving && <Spinner />}
-                </Button>
-              </>
-            ) : (
-              <>
-                {isPaidMeetup && viewing?.payment_status === 'paid' ? (
-                  <Button
-                    variant="destructive"
-                    className="mr-auto"
-                    onClick={() => setRefundConfirmOpen(true)}
-                  >
-                    Refund ticket
-                  </Button>
-                ) : null}
-                <Button variant="outline" onClick={closeDialog}>
-                  Close
-                </Button>
-              </>
-            )}
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
-
-      <Dialog open={refundConfirmOpen} onOpenChange={setRefundConfirmOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Refund this ticket?</DialogTitle>
-            <DialogDescription>
-              {viewing?.ticket_holder_display_name} will be refunded in full and
-              their spot released. Their ticket is kept but marked refunded.
-              This can&apos;t be undone.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button
-              variant="outline"
-              onClick={() => setRefundConfirmOpen(false)}
-            >
-              Cancel
-            </Button>
-            <Button
-              variant="destructive"
-              disabled={isRefunding}
-              onClick={handleRefund}
-            >
-              Refund
-              {isRefunding && <Spinner />}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+        isPaidMeetup={isPaidMeetup}
+      />
     </div>
   );
 };
